pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canctl"
version = "0.1.0"
description = "CAN Bus CLI for LLM — JSON protocol runtime"
requires-python = ">=3.11"
dependencies = [
    "python-can[pcan,vector,kvaser]>=4.0",
    "cantools>=39.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.0",
]

[project.scripts]
canctl = "canctl.canctl:main"

[tool.setuptools.packages.find]
include = ["canctl*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30
